feat: gcs connection 3
